+++ src/software/thunderscope/binary_context_managers/simulator.py
@@ -8,7 +8,7 @@ from proto.import_all_protos import *
 from software.py_constants import *
 from software.thunderscope.proto_unix_io import ProtoUnixIO
 from extlibs.er_force_sim.src.protobuf.world_pb2 import SimulatorState
-from software.thunderscope.binary_context_managers.binary_context_managers import *
+from software.thunderscope.binary_context_managers.util import *
 
 
 class Simulator(object):
