+++ src/software/thunderscope/binary_context_managers/util.py
